# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "200":
          description: User created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{uid}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by Firebase UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (role change, etc)
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks for a user
      parameters:
        - name: uid
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (complete/edit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        name:
          type: string
        email:
          type: string
        grade:
          type: integer
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
      required: [id, uid, name, email, grade, role, createdAt]
    CreateUserBody:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
        email:
          type: string
        grade:
          type: integer
        role:
          type: string
          enum: [user, admin]
      required: [uid, name, email, grade]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        grade:
          type: integer
        role:
          type: string
          enum: [user, admin]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        text:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
      required: [id, uid, text, completed, createdAt]
    CreateTaskBody:
      type: object
      properties:
        uid:
          type: string
        text:
          type: string
      required: [uid, text]
    UpdateTaskBody:
      type: object
      properties:
        text:
          type: string
        completed:
          type: boolean
...